bikesharedata: Pittsburgh Bike Share trips for Q1 of 2021

A dataset containing records of all Pittsburgh Bike Share trips for Q1 of 2021 including start/end times, durations, origins and destinations, and user types.

data(bikesharedata)

Format

A data frame with 14,619 cases of 10 variables:

trip.id

ID Number for the individual trip

Starttime

The date/time the bikeshare trip began

Stoptime

The date/time the bikeshare trip ended

Bikeid

The BikeShare bike's identifier number

Tripduration

The length of the trip in minutes

From.station.id

The origin hub ID number

From.station.name

The origin hub name/location

To.station.id

The destination hub ID number

To.station.name

The destination hub name/location

Usertype

Whether the rider is a customer or subscriber

Source

Adam S Carroll & Pittsburgh Healthy Ride BikeShare, (https://www.kaggle.com/adamscarroll/pittsburgh-healthy-ride-bikeshare-rentals-2021-q1/metadata)

Examples

summary(bikesharedata)
#>     Trip.id           Starttime           Stoptime             Bikeid     
#>  Min.   :111187462   Length:14619       Length:14619       Min.   :70004  
#>  1st Qu.:112065878   Class :character   Class :character   1st Qu.:70540  
#>  Median :112830312   Mode  :character   Mode  :character   Median :70743  
#>  Mean   :112639404                                         Mean   :70708  
#>  3rd Qu.:113190702                                         3rd Qu.:70928  
#>  Max.   :113644504                                         Max.   :71234  
#>                                                                           
#>   Tripduration      From.station.id From.station.name  To.station.id  
#>  Min.   :    60.0   Min.   : 1000   Length:14619       Min.   : 1000  
#>  1st Qu.:   535.5   1st Qu.: 1024   Class :character   1st Qu.: 1023  
#>  Median :  1092.0   Median : 1047   Mode  :character   Median : 1045  
#>  Mean   :  4791.3   Mean   : 9941                      Mean   : 9288  
#>  3rd Qu.:  2931.5   3rd Qu.: 1093                      3rd Qu.: 1093  
#>  Max.   :161392.0   Max.   :88887                      Max.   :88886  
#>                     NA's   :832                        NA's   :966    
#>  To.station.name      Usertype        
#>  Length:14619       Length:14619      
#>  Class :character   Class :character  
#>  Mode  :character   Mode  :character  
#>                                       
#>                                       
#>                                       
#>